class definitions: Observable, Equation
